requirements.txt
flask
pymupdf
groq
python-dotenv
openpyxl